requirements.txt
# requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.7
pillow==10.2.0
torch==2.2.0
numpy==1.26.3
easyocr==1.7.1
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.2
pydantic==2.6.3
pydantic-core==2.16.3
email-validator==2.1.0
python-jose==3.3.0
motor==3.3.1
pymongo==4.6.1
python-dotenv==1.0.1
math==0.0.1